+++ gazebo/Server.cc
@@ -280,11 +280,6 @@ bool Server::ParseArgs(int _argc, char **_argv)
     }
   }
 
-  if (this->dataPtr->vm.count("pause"))
-    this->dataPtr->params["pause"] = "true";
-  else
-    this->dataPtr->params["pause"] = "false";
-
   if (!this->PreLoad())
   {
     gzerr << "Unable to load gazebo\n";
@@ -589,36 +584,14 @@ void Server::Run()
 /////////////////////////////////////////////////
 void Server::ProcessParams()
 {
+  bool p = this->dataPtr->vm.count("pause") > 0;
+  physics::pause_worlds(p);
   common::StrStr_M::const_iterator iter;
   for (iter = this->dataPtr->params.begin();
        iter != this->dataPtr->params.end();
        ++iter)
   {
-    if (iter->first == "pause")
-    {
-      bool p = false;
-      try
-      {
-        p = boost::lexical_cast<bool>(iter->second);
-      }
-      catch(...)
-      {
-        // Unable to convert via lexical_cast, so try "true/false" string
-        std::string str = iter->second;
-        boost::to_lower(str);
-
-        if (str == "true")
-          p = true;
-        else if (str == "false")
-          p = false;
-        else
-          gzerr << "Invalid param value[" << iter->first << ":"
-                << iter->second << "]\n";
-      }
-
-      physics::pause_worlds(p);
-    }
-    else if (iter->first == "record")
+    if (iter->first == "record")
     {
       util::LogRecordParams params;