+++ gazebo/rendering/Conversions.cc
@@ -32,13 +32,7 @@ Ogre::ColourValue Conversions::Convert(const ignition::math::Color &_color)
 }
 
 //////////////////////////////////////////////////
-common::Color Conversions::Convert(const Ogre::ColourValue &_clr)
-{
-  return common::Color(_clr.r, _clr.g, _clr.b, _clr.a);
-}
-
-//////////////////////////////////////////////////
-ignition::math::Color Conversions::ConvertIgn(const Ogre::ColourValue &_clr)
+ignition::math::Color Conversions::Convert(const Ogre::ColourValue &_clr)
 {
   return ignition::math::Color(_clr.r, _clr.g, _clr.b, _clr.a);
 }